Factored out the Atari 2600's 6532 connection, as a low-hanging fruit.
